refactor: 디렉토리 구조 리팩토링
- page/ 하위 구조를 디렉토리로 변경
