+++ static/js/index.js
@@ -1,10 +1,9 @@
 import { preventHyperLinks } from "./tools/preventHyperLinks.js";
 
-import { bindSubMenuEvent } from "./page/subMenu.js";
-import { bindCaroulselEvent } from "./page/carousel.js";
-import { bindWeekdayMenuEvent } from "./page/weekday.js";
-
-import { renderHome } from "./page/home.js";
+import { bindSubMenuEvent } from "./page/sub-menu/bindEvent.js";
+import { bindCaroulselEvent } from "./page/carousel/carousel.js";
+import { bindWeekdayMenuEvent } from "./page/week-day/bindEvent.js";
+import { renderHome } from "./page/home/render.js";
 
 const bindEventListeners = () => {
     bindSubMenuEvent();
